Updated client dashboard to include company creation dialog
